+++ pp/component_from/__init__.py
@@ -1,6 +1,7 @@
 from pp.component_from.gds import gds
+from pp.component_from.gdspaths import gdsdir, gdspaths
 from pp.component_from.netlist import netlist
 from pp.component_from.phidl import phidl
 from pp.component_from.picwriter import picwriter
 
-__all__ = ["phidl", "picwriter", "gds", "netlist"]
+__all__ = ["phidl", "picwriter", "gds", "netlist", "gdspaths", "gdsdir"]